refactor: change file name from layoutheader to layoutbar
